Add bitstring and vector representation functions.
Output contains now also a list of tuples with model information. Tuples
are:

( list of vectors, list of integers, list of bitstrings )

Vectors are simple (x, y, z). Integers are according the Brickodu  rules.
Bitstrings are created from the integers. So each element of the tuple is
a representation of the same model.